requirements.txt
Flask
selenium
python-dotenv